refactor: organize routes with protected group structure
- Move all protected routes into (protected) group folder
- Add login screen outside protected area
- Update root layout to handle protected routes
- Update index.tsx to redirect to protected tabs
- Update video-consumer submodule to latest commit

This improves route organization and prepares for authentication flow.
