+++ app/(protected)/(camera)/_layout.tsx
