+++ app/(protected)/(camera)/drafts.tsx
