+++ app/(protected)/(camera)/post.tsx
