+++ app/(protected)/(camera)/shorts.tsx
