+++ app/(protected)/(tabs)/index.tsx
