+++ app/(protected)/(tabs)/shorts.tsx
