+++ app/(protected)/(tabs)/subscriptions.tsx
