+++ app/(protected)/_layout.tsx
@@ -1,0 +1,105 @@
+import {
+  DarkTheme,
+  DefaultTheme,
+  ThemeProvider,
+} from "@react-navigation/native";
+import { useFonts } from "expo-font";
+import { Stack } from "expo-router";
+import { StatusBar } from "expo-status-bar";
+import { GestureHandlerRootView } from "react-native-gesture-handler";
+import "react-native-reanimated";
+
+import { Redirect } from "expo-router";
+import { PermissionMonitor } from "@/components/PermissionMonitor";
+import { useColorScheme } from "@/hooks/useColorScheme";
+
+export default function ProtectedLayout() {
+  const colorScheme = useColorScheme();
+  const [loaded] = useFonts({
+    "Roboto-Regular": require("../../assets/fonts/Roboto-Regular.ttf"),
+    "Roboto-Bold": require("../../assets/fonts/Roboto-Bold.ttf"),
+  });
+
+  if (!loaded) {
+    // Async font loading only occurs in development.
+    return null;
+  }
+
+  const isLoggedIn = true;
+
+  if (!isLoggedIn) {
+    return <Redirect href="/login" />;
+  }
+
+  return (
+    <GestureHandlerRootView style={{ flex: 1 }}>
+      <ThemeProvider value={colorScheme === "dark" ? DarkTheme : DefaultTheme}>
+        <Stack
+          screenOptions={{
+            headerShown: false,
+          }}
+        >
+          <Stack.Screen
+            name="(tabs)"
+            options={{
+              headerShown: false,
+              presentation: "fullScreenModal",
+              animation: "none",
+            }}
+          />
+          <Stack.Screen name="+not-found" />
+          <Stack.Screen
+            name="onboarding"
+            options={{
+              headerShown: false,
+              presentation: "fullScreenModal",
+              animation: "fade",
+            }}
+          />
+          <Stack.Screen
+            name="(camera)"
+            options={{
+              headerShown: false,
+              presentation: "fullScreenModal",
+              animation: "slide_from_bottom",
+            }}
+          />
+          <Stack.Screen
+            name="preview"
+            options={{
+              headerShown: false,
+              presentation: "fullScreenModal",
+              animation: "none",
+            }}
+          />
+          <Stack.Screen
+            name="reordersegments"
+            options={{
+              headerShown: false,
+              presentation: "fullScreenModal",
+              animation: "none",
+            }}
+          />
+          <Stack.Screen
+            name="upload"
+            options={{
+              headerShown: false,
+              presentation: "fullScreenModal",
+              animation: "fade",
+            }}
+          />
+          <Stack.Screen
+            name="merged-video"
+            options={{
+              headerShown: false,
+              presentation: "fullScreenModal",
+              animation: "none",
+            }}
+          />
+        </Stack>
+        <StatusBar style="auto" />
+        <PermissionMonitor />
+      </ThemeProvider>
+    </GestureHandlerRootView>
+  );
+}
