+++ app/(protected)/onboarding.tsx
