+++ app/index.tsx
@@ -17,7 +17,9 @@ export default function Index() {
   // Handle upload mode with UUID validation
   if (params.mode === "upload") {
     if (params.draftId && isUUIDv4(params.draftId)) {
-      return <Redirect href={`/upload?draftId=${params.draftId}`} />;
+      return (
+        <Redirect href={`/(protected)/upload?draftId=${params.draftId}`} />
+      );
     } else {
       // Could redirect to upload screen without draftId for new recording
       // return <Redirect href="/upload" />;
@@ -25,5 +27,5 @@ export default function Index() {
   }
 
   // Default to tabs
-  return <Redirect href="/(tabs)" />;
+  return <Redirect href="/(protected)/(tabs)" />;
 }
