+++ app/login.tsx
@@ -1,0 +1,9 @@
+import { Text, View } from "react-native";
+
+export default function LoginScreen() {
+  return (
+    <View>
+      <Text>Login</Text>
+    </View>
+  );
+}
